folder tree structured, sass implemented
