+++ src/app/dashboard/transfer-history/transfer-history.component.scss
@@ -1,0 +1,37 @@
+@import '../../scss/vars';
+
+.search {
+  margin-top: 83px;
+
+  .form_title {
+    padding: 10px;
+    font-size: 1.2rem;
+    color: $placeholder-color;
+  }
+}
+
+.search-input {
+  max-width: 170px;
+}
+
+
+.history__table--sticky-header {
+  min-height: 25vh;
+  overflow-y: auto;
+  height: 25vh;
+  bottom: 200px;
+
+  thead th {
+    position: sticky;
+    top: 0;
+  }
+
+  @media only screen and (max-width: 800px) {
+
+    th {
+      min-width: 124px;
+    }
+  }
+
+
+}
