+++ src/app/forms/new-jar-form/new-jar-form.component.ts
@@ -1,31 +1,27 @@
 import { Component, OnInit, ViewChildren, ElementRef, AfterViewInit } from '@angular/core';
 import { FormControlName, FormGroup, FormBuilder, Validators } from '@angular/forms';
-import { IJar } from '../jar/jar-interface';
+import { IJar } from '../../jar/jar-interface';
 import { Subscription, Observable, fromEvent, merge } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
-import { JarService } from '../services/jar.service';
-import { CurrencyService } from '../services/currency.service';
-import { debounceTime } from 'rxjs/operators';;
+import { JarService } from '../../services/jar.service';
+import { debounceTime } from 'rxjs/operators';
+import { CurrencyService } from '../../services/currency.service';
 
 @Component({
-  selector: 'app-currency-settings',
-  templateUrl: './currency-settings.component.html',
+  selector: 'app-new-jar-form',
+  templateUrl: './new-jar-form.component.html'
 })
-export class CurrencySettingsComponent implements OnInit, AfterViewInit {
+export class NewJarFormComponent implements OnInit, AfterViewInit {
   @ViewChildren(FormControlName, { read: ElementRef }) formInputElements: ElementRef[];
 
-  pageTitle = 'Currency Settings';
+  pageTitle = 'Create New Jar';
   errorMessage: string;
   erronOnCreatedJar: string;
-
-  newSettingsForm: FormGroup;
-  createdSettingsForm: IJar;
+  newJarForm: FormGroup;
 
   jars: IJar[] = [];
+  createdJarForm: IJar;
   currencyOptions: string[];
-  currencyOptions1: [];
-  private sub: Subscription;
-
 
   constructor(private fb: FormBuilder,
     private route: ActivatedRoute,
@@ -33,14 +29,15 @@ export class CurrencySettingsComponent implements OnInit, AfterViewInit {
     private jarService: JarService,
     private currencyService: CurrencyService) { }
 
-
   ngOnInit(): void {
-    this.newSettingsForm = this.fb.group({
-      jarName: ['', Validators.required],
+    this.newJarForm = this.fb.group({
+      jarName: ['', [Validators.required,
+      Validators.minLength(3),
+      Validators.maxLength(25)]],
+      accountBalance: ['', [Validators.required, Validators.min(0), Validators.max(100000)]],
       currency: ['', Validators.required],
 
     });
-
 
     this.jarService.getJars()
       .subscribe({
@@ -54,7 +51,6 @@ export class CurrencySettingsComponent implements OnInit, AfterViewInit {
       .subscribe({
         next: data => {
           this.currencyOptions = data;
-          this.currencyOptions1 = data;
         },
         error: err => this.errorMessage = err
       });
@@ -63,32 +59,20 @@ export class CurrencySettingsComponent implements OnInit, AfterViewInit {
   ngAfterViewInit(): void {
     const controlBlurs: Observable<any>[] = this.formInputElements
       .map((formControl: ElementRef) => fromEvent(formControl.nativeElement, 'blur'));
-    merge(this.newSettingsForm.valueChanges, ...controlBlurs).pipe(
+    merge(this.newJarForm.valueChanges, ...controlBlurs).pipe(
       debounceTime(800)
     );
   }
 
-  onSettingsChange() {
-    for (const jar of this.jars) {
-      if (this.newSettingsForm.value.jarName.id === jar.id) {
-        jar.currency = this.newSettingsForm.value.currency;
-        this.createdSettingsForm = { ...this.newSettingsForm.value };
-        this.createdSettingsForm = jar;
-      }
-    }
-  }
-
-
-  submitSettings(): void {
-    if (this.newSettingsForm.valid) {
-      if (this.newSettingsForm.dirty) {
-        this.onSettingsChange();
-        this.jarService.updateJar(this.createdSettingsForm)
+  submitNewJar(): void {
+    if (this.newJarForm.valid) {
+      if (this.newJarForm.dirty) {
+        this.createdJarForm = { ...this.newJarForm.value };
+        this.jarService.addNewJar(this.createdJarForm)
           .subscribe({
+            next: () => { },
             error: err => this.erronOnCreatedJar = err
           });
-      } else {
-        this.onJarCreated();
       }
     } else {
       this.errorMessage = 'Please correct the validation errors.';
@@ -97,7 +81,7 @@ export class CurrencySettingsComponent implements OnInit, AfterViewInit {
   }
 
   onJarCreated(): void {
-    this.newSettingsForm.reset();
+    this.newJarForm.reset();
     this.router.navigate(['/']);
   }
 }
