+++ src/app/navbar/navbar.component.html
@@ -1,3 +1,4 @@
+
 <nav class="navbar navbar-expand-md navbar-dark bg-secondary fixed-top">
   <div class="container">
     <a class="navbar-brand text-warning" (click)="isCollapsed = !isCollapsed" [routerLink]="['/']">MoneyBox App</a>
